requirements.txt
pyyaml==6.0.1
ruff==0.1.8
black==23.12.0